Add files via upload
Hi everyone,
This is my autocorrelation study on the data.  Conclusions are at the bottom of the notebook.  See if you have any other or different  insights after looking at the data.

A few take-aways' about big data processing.  I first tried this notebook on a small sample of data using our scrubbed dataset and it worked well.  When I increased the data slice to the whole set things went badly and the kernel died - long live the kernel!
The technique I applied to avoid slow or failed processing, which greatly speed up processing, was to only work with the most minimal data set I required by not importing or by del as many columns as I could. 
Even having a pickup-time column remain after I used it to create an index slowed things down a lot.
If you look at the def, you'll see that I delete the sampled database each time to keep a small memory profile. 
I found it is better to reload data then reuse data in some cases.